Fixed bug #3561047 (http://bit.ly/Pf5L5v): Category overview sometimes doesn't start per preferences.
Main form was indirectly changing the tab in the overview pane when starting up with any except Categories tab (tab index 0) selected. This was saving state of tab 0 before it was assigned any tree nodes, meaning that it was displayed expanded when next selected because of lack of state info. Fix meant providing any Initialise method for the overview pane that selects the required tab without saving the state of tab 0. Main form calls this via new Initialise method of TMainDisplayMgr.