Added JavaScript fix for auto overflow bug in IE that is optionally inserted into detail.html as required. The new script is stored in overflowXFix.js which is embedded in HTML resources.
This fixes display problem where an unexpected vertical scroll bar appears when an element overflows horizontally when overflow-x CSS property is set to "auto".

Changed TDetailPageLoader so that HTML entity conversions are not performed on scripts.

Corrected body of detail.html to be enclosed in a div instead of a paragraph.