Merged revision(s) 3581-3683, 3314-3582 from trunk:
Updated with changes in trunk/Src after applying changes from release 4.8.2 to it.

Resolved conflicts with .vi files by ignoring merged changes.
Changes to PJSysInfo.pas in trunk were not merged because that unit is in a different location in parsnip.
Changes to UFontHelper.pas in trunk were identical to changes already made in parsnip, so the file was not changed in parsnip.

The outcome of the merge was simply to update the v4 change log.