Reverse merged revision(s) 3836-3835 from branches/parsnip/Src/Main/CS.Markup.pas:
This restores TMarkup to a record instead of a class and interface.

I've decided to keep TMarkup as a dumb receptacle for markup at the moment: any conversion to other types will be done elsewhere. This is because there are various different ways of interpreting plain text markup in different parts of the program and different ways of converting REML to plain text (usually via Active Text), so it seems best to leave interpretation to the parts of the program where markup is interpreted.