Revised SWAG importing to use local data instead of web service.
Local data is a folder containing a copy of the SWAG source from the delphidabbler/swag project on GitHub. It includes a main XML file and separate source files for each snippet.
As few changes made as possible. A new SWAG.UXMLProcessor unit provides access to the XML file
SWAG.UReader was modified to use SWAG.UXMLProcessor instead of Web.USWAGRESTMgr to get SWAG data,
A new tab was added to TSWAGImportDlg where the user provides the path to the local folder containing SWAG data. The form was also modified to validate the data path and to use the modified data format.