Permit hiding restart message in printer prefs
When the printer preferences page of the preferences dialogue box is
displayed as part of the full dialogue box it displays a message to
inform that the settings won't take place until the program is
restarted. But we don't want that message when frame is displayed on
its own when displayed from print dialogue box. The new mechanism to
pass custom flags to frames is used to pass a flag that hides the
message.

Also the message label was changed from static text to a TLabel to
enable it to be displayed in the program's warning colour.