Comment out call to set theme on theme change.
This was necessary because the web browser control throws a peculiar
exception if the theme is changed while it is displayed.

Because of this it has been decided that CodeSnip won't change its theme
in response to the user changing it in Windows settings. Instead it will
have a choice of themes that will be active until explicitly changed
(like Delphi 11).