Update TCSSSelector for fluid chaining of methods
Modify TCSSSelector.AddProperty to reference to current TCSSSelector
instance to permit chaining the method to others that return aborts
TCSSSelector reference.

Add new TCSSSelector.AddPropertyIf fluid method to permit a choice of
two different properties to be added depending on the value of a Boolean
variable. This was created prevent chaining from being broken when a
conditional statement was encountered.

Closes #73