Add FmRegisterCompilersDlg dialogue box to project.
This dialogue box is designed to present a list of Delphi compiler(s) to
the user to enabled them to select which, if any, compiler(s) are to be
registered.

Designed for display at startup when unregistered versions of Delphi are
detected.

The dialogue includes an HTML resource that is used to display help
information on the face of the dialogue box.