Add support for rendering REML lists as plain text
ActiveText.UTextRenderer has support added to parsing and rendering REML
lists and to format the resulting text with given margins and page
widths. This was provided in a new TActiveTextTextRenderer.RenderWrapped
method.

TActiveTextTextRenderer.Render was made private since it is now only
called internally.

USourceGen & UTextSnippetDoc were modified to use the above new method.

** This solution is a little kludgy and relies on parsing NBSP
characters emitted by TActiveTextTextRenderer.Render. This was done
originally because the formatting was being done in USourceGen and
UTextSnippetDoc which didn't have access to the inner workings of
TActiveTextTextRenderer. The formatting code was then moved into
TActiveTextTextRenderer and so there's probably a more elegant
solutuion available now.