Missing text at end of detail pane list items fixed
Code added to track and detect if text is being written inside a list
item and to output any inline text as HTML between the end of a nested
list and the end of a list item.

Such detection is then used in logic that determines when to output such
orphaned text.

Fixes #82