Major overhaul of active text HTML renderer
Code now based in part on the REML active text writer. The REML based
code was simplified due to the fact that the HTML renderer has tags that
match the ekBlock and ekDocument elements, whereas REML doesn't.

HTML code is now indented to highlight the document structure (although
this doesn't matter really since the code is never displayed to the
user).