Prevent snippets editor from stripping REML p tags
If a single paragraph was entered in the snippet's editor's REML editor
in form <p>content here</p>, with no inline tags the <p> tags would be
stripped. This is undesirable because text outside of <p> tags may be
rendered differently to text in <p> tags, so stripping the tags could
change the rendering.

This behaviour is changed by this commit and <p> tags are now preserved.