Add new DB.Collections unit
Provides support for multiple snippets collections.

Contains code to:

* Encapsulate a collection ID
* Encapsulate a collection location
* Encapsulate a collection, including ID, name, location and file format.
* Maintain a single list of available collections.
* Persist the collection list to settings.

Uses a singleton to maintain the list of collections.

Ensures that collections are always defined that are the equivalent of
the CodeSnip 4 "main" & "user" databases. This code may be removed when
the changeover from the two databases to multiple data sources is
achieved.

Some __TMP__ methods are provided to ease the changeover from
CodeSnip v4 to v5.