Remove IView.IsUserDefined method
IView.IsUserDefined was removed along with all its implementations.

The only place that IView.IsUserDefined was used was in
FrOverview.TOverviewFrame.TTVDraw.IsUserDefinedNode. This method was
modified to retrieve any collection ID associated with any snippet or
category view and to calculate its result using those, removing any
need for IView to provide a IsUserDefined method.

A __TMP__ method was required in the modified TTVDraw.IsUserDefinedNode
method.

Removed code was commented out.