Change interface used to set collection colour prefs
Changed IPreferences and its implementation to remove DBHeadingColours[]
property & getter and setter that took Boolean (user-defined) parameter
and to replace with new Getter & Setter methods that work with
collections IDS.

Made similar changes re DBHeadingCustomColours[] property & getter and
setter.

The new setter and getter methods for both above still access the
existing fDBHeadingColours[] and fDBHeadingCustomColours[] fields which
remain array[Boolean] types, which must be changed to support a variable
number of collections.

Updated TPreferencesPersist to use the new TCollectionID version of the
getters and setter methods, while retaining the CodeSnip 4 setting
keys (something else that needs changing when moving to a variable
number of collections).

Made changes to all code that uses the removed properties to work with
the new getters and setters.

Several __TMP__ method calls were required to make this work, although
a few were able to be removed.

Introduced new __TMP__ method within UPreferences to map collection IDs
to the colour preferences for the CodeSnip 4 main and user databases.

Removed code was commented out.