Store collection IDs in FmDependenciesDlg's listbox
FmDependenciesDlg used to record TBox<Boolean> objects in the Objects[]
array properties of the list box that displays dependent snippets. This
was used to set the colour of a snippet depending on whether it was user
defined or not. This was changed to store a TBox<TCollectionID> object
instead of the boxed Boolean object so that collection ID can be used to
get the required snippet colour from preferences.

Eliminated a __TMP__ method call.

Removed code was commented out.