Rename DBIO.UIniDataReader unit as DBIO.UIniData
Renamed the unit since it is going to be used for writing the Code
Snippets Collection ini database format (known as the "main database" in
CodeSnip 4) in addition to the existing code used to read the
collection.

Modified the CodeSnip .dpr and .dproj project files re the name change.

Updated use clause in DB.UDatabaseIO re the name change.