Add new TDatabase.GetUniqueSnippetKey method
This method was added to IDatabaseEdit and implemented in TDatabase.

It returns a 32 character alphabetical value that is unique within a
given snippet collection.

Its purpose to for use in new snippet keys.

The key is also a valid Pascal identifier, so can be used as a unit name
for test compiling.