Auto-generate snippet key in duplicate snippets dlg
Removed the ability for the user to specify the snippet key (formerly
the snippet name) in the Duplicate snippet dialogue box. The new snippet
key is now automatically generated and is certain to be unique within
the snippet's collection.

It is now compulsory to enter a display name (because the key that is
used for the display name by default is now a pseudo random sequence of
characters & makes no sense as a name).

Hard wired the "user" snippets collection. Later this should be chosen
by the user.

Put snippet display name in dialogue box caption in double quotes.