Remove editing of snippet key from Snippets Editor
Removed the ability for the user to specify the snippet key (formerly
the snippet name) in the Snippets Editor dialogue box. When creating new
snippets, the new snippet key is now automatically generated and is
certain to be unique within the snippet's collection. When editing
existing snippets the original snippet ID is preserved.

It is now compulsory to enter a display name, for two reasons:

1. because the key that is used for the display name by default is now a
pseudo random sequence of characters & makes no sense as a name.

2. because the key can no longer be edited.

Hard wired the "user" snippets collection for new snippets. Later this
should be chosen by the user. Existing snippets being edited can't be
moved from their original collection.

Put snippet display name in dialogue box caption in double quotes.