Change how SWAG import handles snippet keys
The snippet key under which a SWAG snippet is to be created has been
changed. It was previously created from concatenating "SWAG" with the
SWAG ID number of the snippet. It is now generated by calling
IDatabaseEdit.GetUniqueSnippetKey, which brings it into line with how
other parts of CodeSnip create new keys. As part of this change, the ID
of the collection receiving the snippets was required.

A method was added to TSWAGImportDlg to get the ID of the collection in
which to add imported SWAG snippets. This method is a stub that always
returns the "user" collection ID. In later updates it will be changed to
get the required collection from the user.

The SWAG import wizard no longer displays the key to be used for
imported snippets.