Add collection ID param to TDatabase.CreateTempSnippet
Updated TDatabase.CreateTempSnippet to take an additional TCollectionID
parameter and to use that as the temporary snippet's collection ID
instead of always using the "user" collection.

Renamed some the parameters.

Updated affected calling code:

* Updated one of the TSnippetValidator.ValidateDependsList overloaded
methods to also add a collection ID parameter so that it can pass it
along to the revised TDatabase.CreateTempSnippet method.

* Updated FmSnippetsEditorDlg to pass the required collection ID
parameter to both TDatabase.CreateTempSnippet and
TSnippetValidator.ValidateDependsList.