Ensure a default category is always created
A new TCategoryEx.CreateDefault method was added that creates a default
category with ID '__default__' and description "My Snippets".

TDatabase.Load now call the above method to create a default category if
one with the same ID was not loaded. This new category can be deleted,
but it will always be recreated. The category can, though, be renamed.

Updated FmSnippetsEditorDlg to use the new category as the default
selection in the snippets editor dialogue box.