Fix bug in DBIO.UIniData
If the "main" collection is not loaded first then any categories that
have already been loaded that it doesn't recognise would cause a crash
when the loader tried to open a non-existent category .ini fle.

Added an if statement to check for the existance of the required .ini
file before trying to open it.