Change native file format to ignore empty categories
The "native" CodeSnip 4 file format (formerly the user database format)
no longer saves details of empty categories.

Because TNativeV4FormatSaver now behaves exactly the same as
TDCSCV2FormatSaver in not writing empty categories, the code that
ensures this was pushed up into the parent class and a virtual abstract
method and its overrides were able to be removed.