Change branding, directories & version for CS Vault
Changed per-user and global application directories from CodeSnip.4
to CodeSnip.Vault to avoid changes to file formats when developing
Vault from destroying files required to run existing CodeSnip 4
installations.

Changed program name to CodeSnip.Vault to make it easy to see that the
development version of CodeSnip vault is being run. This change also
made in version information.

Changed version number to 4.900.0 and build number to 1000.

Changed splash screen to one with "Vault" overlaid on it.