Enable snippets from "main database" to be edited
Several files were altered in order to do this:

* ExternalObj.ridl, external.js & UWBExternal.pas were all modified so
that the external object's EditSnippet method could take an additional
parameter containing a hex representation of a collection ID so that the
Edit Snippet link in the Details pane would work for > 1 collection.

* IntfNotifier & UNotifiers had their EditSnippet method modfied to
take a second collection ID parameter.

* UUserDBMgr had several sections of code altered to (a) permit "main"
snippets to be edited and (b) to handle snippet collections.

* TEditSnippetAction was overhauled to expose a snippet ID property
instead of a Key property to enable both the snippet key and collection
ID to be passed along.

* DB.UMain was modified to permit "main" snippets to be edited.

* UDetailPageHTML, FmMain and FmDuplicateSnippetDlg were all modified to
work with the code changes above.