Revised Favourites file format to record collections
The Favourites file format was redefined to be able to store a snippet's
collection ID in hex format, along with the snippet's key and the date
the favourite was last viewed.

The date format used to store the last viewed date was changed from
local to international YYYY-MM-DD HH:MM:DD format.

The saving and loading code in TFavouritesPersist was rewritten to use
the TTabSeparatedReader and TTabSeparatedFileWriter classes from
UTabSeparatedFileIO to perform the reading, low level parsing, and
writing of the favourites file.

A new TAppInfo.UserFavouritesFileName method was added to return the
full path to the current user's favourites file, meaning that the file
name is no longer hard wired in TFavouritesPersist.