Revised snippet selection file format re collections
The Snippet Selection file format was redefined to be able to store a
snippet's collection ID in hex format, along with the snippet's key.

The saving and loading code in TSnippetIDListFileReader and
TSnippetIDListFileWriter was rewritten to use the TTabSeparatedReader
and TTabSeparatedFileWriter classes from UTabSeparatedFileIO to perform
the reading, low level parsing, and writing of file.

The file watermark was updated to v2.