Revise display preferences re collections
Updated display preferences stored in config file by removing settings
for the old "main" and "user" database snippet colours and replacing
them with separate settings for snippet heading colours for each
collection and a different colour setting for group headings.

Heavily revised the Display page of the Preferences dialogue box to
allow the user to enter and edit the new colour preferences.

Replaced default colours for "main" and "user" snippets with default
colours for snippet and group headings.

Made minimal changes to code that uses the changed preferences to
display snippet headings and tree nodes in the colour assigned to their
collection and/or to display group headings in the colour assigned to
them. It most cases preferences method calls on one or two lines were
all that was required. In other cases more significant changes were
required.