Modify About box database tab to show collection info
Replaced the About box's "About The Database" tab with an "About
Collections" tab that asks user to choose a collection and then shows
any meta data associated with the collection. The data is shown, grouped
by type, in a tree view. Where the collection has no meta data, a
message for that is shown.

Redundant code was deleted and there was some renaming for clarity.

The dlg-about-database-tplt.html HTML template file used the the "About
The Database" tab was deleted and its reference was removed from
HTML.hrc.