Modify About Box to list collection paths
The Paths & Files tab of the About box was revised to list the paths of
each collection's data instead of the paths to the "user" and "main"
databases.

The new design places the custom group boxes used to display the path
information within a scroll box since the number of collections can't be
predicted and it can't be guaranteed that the path group boxes would not
overflow the availablr screen space.

Moved the buttons used to display the config files from the bottom to
the top of the Paths & Files tab.

Updated commenting to reflect the changes and to fix some typos.