Remove "user database" checks from UCodeShareMgr
TCodeShareMgr was requiring that snippets were available in the "user
database" before permitting code to be exported. This was changed to a
simple check that the entire database is non-empty.

TCodeShareMgr was also refusing to extract a snippet object from a
snippetview if the snippet didn't come from the "user database". This
check was eliminated.

Code comments referring to the "user database" were edited. All class
and method comments were converted to XMLDoc format.