Update code importer snippet dependency rules.
Previously UCodeImportMgr checked whether any snippets in an imported
snippet's depends list were in either the user or main databases and
deleted them if not found in either.

The importer now requires that imported snippets can only list snippets
in their depends list if those snippets are also in the import. Any
snippets that are not in the import are removed from depends lists.

This was required to remove code that required the DCS collection to be
present, which can no longer be guaranteed. Additionally, it's
dangerous to assume that any snippet outside the import whose key is
referenced from the import is the same snippet as was intended.

All imported snippets are now given a new key within the chosen
collection.

UCodeImportMgr had to be heavily overhauled to get this new system to
work. In particular the use of new keys meant that the depends lists
needed to be fixed up with new keys.