Revise snippet exporter re changes in importer
TCodeExporter was modified to not include the keys of snippets in
depends lists for snippets not themselved included in the export.

Also fixed an error where the old snippet ID for snippets included in
the export was being used in depend lists. This meant that no snippet
depends lists were ever imported.
TCodeExporter was modified to ensure that exported dependency lists use
the same key as the changed snippet key.