Replace update action testing for nonempty user db
The ActNonEmptyUserDBUpdate action update event handler that tested for
a non-empty user database was removed and calls to it were replaced by
calls to ActNonEmptyDBUpdate that makes no such tests.

NOTE: menu actions that used this event handler to enable/disable them
are still "User Database" specific and need to be removed or replaced.