Change Database move code to move any collection
The code in UUserDBMove was modified to move any collection's data files
instead of only the "user database".

FmUserDataPathDlg was also modified to work with the revised
UUserDBMove. In addition, the facility to restore the "user database" to
its default location was removed since there is no concept as a default
location for collections.

Text in FmUserDataPathDlg and UUserDBMove was changed to refer to
collections instead of databases. Comments in both units were similarly
updated.

Text and hints in the relevant menu item in FmMain was changed to refer
to collection instead of user database.