Add new v5 folder backup file format & make default
A new v5 folder backup file format was added. This new format adds a
custom field to the file header. This is a variable length array of
bytes preceeded by a 16 bit byte count. The custom header field can be
from 0 to High(Int16) bytes long. The FileID from previous formats was
retained.

The v5 file format is now the default file used for writing backup
files. Ability to read formats v1 to v4 in addition to v5 has been
retained.

The idea of the new custom header is to allow a byte array
representation of a collection ID to be stored in the file header thus
enabling a check to be made that a backup file contains a specific
collection's data files.