Change Database backup code to backup any collection
The code in UUserDBBackup and UUserDBMgr was modified to backup and
restore a specified collection's data files. Before only the "user
database" was handled.

A new FmCollectionBackup dialogue box was added to the project. This is
used when backing up and restoring to get the collection to be backed up
/ restored and the name of the backup file.

The backup and restore menu items and hints were updated to refer to
collections instead of the user database.

Many comments and method names still refer to the "user" database and
need to be updated.