Update DCSCv2 saver to use new v5 backup file format
The code that saves the DelphiDabbler Code Snippets Collection's content
to storage backs up the collection before saving so that it can be
restored should anything go wrong. This backup / restore code was
modified to call TUserDBBackup instead of directly calling TFolderBackup
so that the correct FileID and Custom header for this file format
backups are written to the backup file.