Change database delete code to work with collections
The code in UUserDBMgr was changed to delete all snippets in a given
collection rather than to delete just the "user database".

The FmDeleteUserDBDlg was modified to not only get permission to delete
the collection but also to get the user to select collection whose
data is to be deleted.

The delete database menu item and hints were updated to refer to
collections instead of the user database.

The explanatory HTML that appears in FmDeleteUserDBDlg was modified re
the change to deleting a selected collection.