Prevent default collection linking to DCSC snippets
TNativeV4FormatLoader was setting snippets dependency and xref
properties to include snippets in the "main database" (now DelphiDabbler
Code Snippets Collection - DCSC). While this was OK when there were just
the "main" and "user" databases, it is not permitted with collections:
we can't assumed anything special about the DCSC, which may not even be
installed.

Therefore snippets in collections loaded by TNativeV4FormatLoader will
have any links to the old "main" database ignored.

This change meant that TNativeV4FormatLoader.FindSnippet was now
implemented identically to TDCSCV2FormatLoader.FindSnippet, so the
implementation was pushed up into the base class and the methods were
removed.