Change database load/save to cycle thru collections
TDatabase's Load and Save methods previously had the "user" and "main"
databases (later DCSC and Default collections) loaded/saved explicitly.

This was changed so that the methods loop through all available
collections, loading and saving them. This means that the methods will
now load and save whatever collections are available.