Update TAppInfo with collection related directories
Added per-user directories TAppInfo.UserCollectionsDir and
TAppInfo.UserDefaultCollectionDir. The former is the parent directory
under which it is recommended that collections are stored. The latter is
the directory where CodeSnip will initially create the Default
collection, which is a sub-directory of TAppInfo.UserCollectionsDir.

TCollections.Initialize was modified to create the Default directory,
when necessary, in TAppInfo.UserDefaultCollectionDir instead of
TAppInfo.DefaultUserDataDir.

TAppInfo.DefaultUserDataDir was removed as it is now redundant.