Fix bug in TMainDBMetaData not finding moved DCSC
When the DelphiDabbler Code Snippets Database is moved from the location
specified by TAppInfo.AppDataDir (which was compulosory in CodeSnip 4),
TMainDBMetaData could no longer read meta data because it was always
looking for it in TAppInfo.AppDataDir.

TMainDBMetaData was changed to look for the data in the location
specified by a TCollection instance passed to to the constructor.