Revise how TCollection handles data storage info
TCollection previously and a Location of type TCollectionLocation
property that stored the directory in which the collection is stored
along with a CollectionFormatKind property to store the collection's
data format. TCollectionLocation also had some other properties that
have never found a used.

A new record, TDataStorageDetails, was added to DB.DataFormats that
records both the data format kind and the directory used to store the
collection.

TCollection replaced its Location and CollectionFormatKind properties
with a single Storage property of type TDataStorageDetails.

The key/value pairs in the [Collection:X] config file sections were
revised to store the new TCollection property values and not to store
the removed property values.

Calling code was modified as necessary re these changes.

The redundant TCollectionLocation record was deleted.