Document CodeSnip Vault native data format
Added new HTML file format document for the CodeSnip Vault native data
format. This file not only documents the features of the format as
currently implemented but also documents how it is proposed to implement
storage of meta data in future. The proposed features are currently
commented out so they don't appear in the rendered HTML document.