Renaming in welcome page
References to collections welcome-tplt.html were changed to refer to
vaults.

HTML "collections" id was changed to "vaults" in welcome-tplt. The
related CSS was changed similarly in detail.css.

Placeholders in welcome-tplt that referred to collections were to
changed to refer to vaults. The resolution of the placeholders in
UDetailPageHTML were changed to match.

Local variables in UDetailPageHTML that referred to collections were
renamed to refer to vaults.