Merge code from UIniDataLoader into DBIO.UIniData
The TDatabaseIniFile class in UIniDataLoader has been reduced to a rump
that is essentially a ini file extension that strips surrounding quotes
from string values read from the ini file.

The entire class was moved to be a nested class of TIniDataReader. It
was renamed as TUTF8IniFileEx. The EDatabaseIniFile exception was not
brought over: EDataIO was used instead.

UIniDataLoader was removed from the project.