Rename DBIO.UIniData as DB.IO.DataFormat.DCSCv2
The old name of this unit was not indicative of the purpose of this
unit. It is specialised for reading / writing the DelphiDabbler Code
Snippets Collection v2 format. The fact that the .ini file format is
used as part of the format is secondary.

The uses statement of DB.UDatabaseIO was updated to reference the
renamed unit.