Refactor out TSnippetIDListEx
TSnippetIDListEx added an extra constructor to TSnippetIDList that
extracted snippet IDs from an TSnippetList instance. This was replaced
by a new TSnippetList.IDs method that returns an ISnippetIDList instance
containing the IDs of all the snippets in the list.

Code that used TSnippetIDListEx was changed to use the TSnippetList.IDs
method instead.