Revise Delete Vault dialogue box
The dialogue box was revised to give two options:

1) To delete all the snippets in a vault but leave an empty vault and data
directory in place.

2) To delete all the snippets in a vault and to delete the vault and remove its
data directory.

The file containing HTML displayed in the dialogue box was rewritten. The HTML
file was renamed and its entry in HTML.hrc modified accordingly.

The caption of the dialogue box was changed to reflect the options available.

TUserDBMgr.DeleteDatabase was modified re the changes in the dialogue box:
DeleteDatabase now implements the deletion options chosen in the dialogue box.

The menu item text and hint in FmMain was altered to reflect the new dual
purpose of the dialogue box.