Fix error in export file format documentation
Documentation of XML tags relating to storing snippet xrefs in export
files was removed as erroneous. The error was noted and the file format
bumped to v8 following this change.

Updated the current file version number in the UCodeImportExport unit to
v8.

Fixes #151
Updated the export help topic to note that snippet categories and xrefs
are not exported.