Revise handling of file extensions in save dialogues
This commit applies to the custom file save dialogues used when saving
units or annotated source code files.

The code to detect file types and to look up filter indexes depended on
file extensions being different for each file type, which cannot be
guaranteed.

Relevant code was updated to remove this dependency.

NOTE: This change was made because HTML 5 documents, when added, will
use the same .html file extension and XTHML documents.