Fix TSaveSourceMgr to preview HTML5 files correctly
HTML 5 files are actually previewed as XHTML because support for HTML5
might not be properly implemented by the IE based MS web browser
control.

TSaveSourceMgr was also changed to explicitly render pascal file as
plain text for previewing. Although this change is not strictly
necessary since pascal and plain text file types render the same, it
was done for consitency and clarity.