Add new THML5 & TXHTML classes & reimplement THTML
THTML was re-implemented as a static abstract class instead of a record.
It now acts as an abstract base template class for two new static
classes - TXHTML and THML5.

TXHTML was added as drop-in replacement for the original implementation
of THTML. Since the original THTML generated XML compliant tags, this
new name is more accurate.

All code that called THML was changed to call TXHTML instead. TXHTML was
implemented without any change of interface to THTML, despite the fact
that the original THML was a record and TXHTML is a class.

THTML5 was added as another static class that descends from THTML. This
class generates HTML 5 compliant tags.

A new class reference to THTML derived types was added, named
THTMLClass.