Fix handling of Unicode chars in RTF generation.
Modified check in WideCharToChar in UEncodings unit to detect failed
conversions from WideChar to a given ANSI code page.

Added support for \uc RTF control word to TRTF in URTFUtils unit.

Fixed problem in TRTF.MakeSafeText that was not handling Unicode
correctly when a Unicode character cannot be represented in a given
code page.

Fixes #157

Also fixed error in detecting ASCII characters that need to be escaped#
in RTF (per issue 159)

Fixes #159