Add TVaults method to generate a truly unique ID
Added new TVaultID.CreateNew method that returns a new UUID.

The new TVaults.GetUniqueID repeatedly calls TVaultID.CreateNew until
a vault ID is returned that doesn't already exist.

Note that since TVaultID.CreateNew calls TUniqueID.NewID the new ID is
a GUID and so should already be unique, but it's safer to have the
ability to guarantee it.